Move settings to connectors package
